requirements.txt
python-dotenv~=1.1.1
pymongo~=4.15.2
motor~=3.7.0
RapidFuzz~=3.14.0
pytz~=2025.2
Unidecode~=1.4.0
tabulate~=0.9.0
uvicorn~=0.37.0
itsdangerous
starlette~=0.48.0
pendulum~=3.1.0
aiohttp~=3.12.15
fuzzywuzzy~=0.18.0
discord.py~=2.6.3
pillow~=11.3.0
requests~=2.32.5
backoff~=2.2.1
psutil>=5.9.0
